Rename split modules to split_tunnel
