+++ mullvad-cli/src/cmds/mod.rs
@@ -35,9 +35,9 @@ mod reset;
 pub use self::reset::Reset;
 
 #[cfg(target_os = "linux")]
-mod split;
+mod split_tunnel;
 #[cfg(target_os = "linux")]
-pub use self::split::Split;
+pub use self::split_tunnel::SplitTunnel;
 
 mod status;
 pub use self::status::Status;
@@ -63,7 +63,7 @@ pub fn get_commands() -> HashMap<&'static str, Box<dyn Command>> {
         Box::new(Relay),
         Box::new(Reset),
         #[cfg(target_os = "linux")]
-        Box::new(Split),
+        Box::new(SplitTunnel),
         Box::new(Status),
         Box::new(Tunnel),
         Box::new(Version),
